+++ src/core/EVM/EVMStepExecutor.ts
@@ -52,8 +52,8 @@ import { isGaslessStep, isRelayerStep } from './typeguards.js'
 import type { Call, WalletCallReceipt } from './types.js'
 import {
   convertExtendedChain,
+  getDomainChainId,
   getMaxPriorityFeePerGas,
-  isSaltMatchingChainId,
 } from './utils.js'
 import { waitForBatchTransactionReceipt } from './waitForBatchTransactionReceipt.js'
 import { waitForRelayedTransactionReceipt } from './waitForRelayedTransactionReceipt.js'
@@ -246,8 +246,14 @@ export class EVMStepExecutor extends BaseStepExecutor {
         id: stepBase.id,
       }
     } else {
-      const params = signedTypedData?.length
-        ? { ...stepBase, typedData: signedTypedData }
+      const filteredSignedTypedData = signedTypedData?.filter(
+        (item) => item.signature
+      )
+      const params = filteredSignedTypedData?.length
+        ? {
+            ...stepBase,
+            typedData: filteredSignedTypedData,
+          }
         : stepBase
       updatedStep = await getStepTransaction(params)
     }
@@ -578,7 +584,8 @@ export class EVMStepExecutor extends BaseStepExecutor {
           if (!this.allowUserInteraction) {
             return step
           }
-          const typedDataChainId = typedData.domain.chainId as number
+          const typedDataChainId =
+            getDomainChainId(typedData.domain) || fromChain.id
           // Switch to the typed data's chain if needed
           const updatedClient = await this.checkClient(
             step,
@@ -626,8 +633,7 @@ export class EVMStepExecutor extends BaseStepExecutor {
         const signedNativePermitTypedData = signedTypedData.find(
           (p) =>
             p.primaryType === 'Permit' &&
-            (p.domain.chainId === fromChain.id ||
-              isSaltMatchingChainId(p.domain.salt as Hex, fromChain.id))
+            getDomainChainId(p.domain) === fromChain.id
         )
         if (signedNativePermitTypedData) {
           transactionRequest.data = encodeNativePermitData(